+++ src/main/kotlin/restApi/Models.kt
@@ -4,11 +4,11 @@ import com.neelkamath.omniChatBackend.db.Pic
 import com.neelkamath.omniChatBackend.graphql.routing.MessageText
 
 data class InvalidMediaMessage(val reason: Reason) {
-    enum class Reason { USER_NOT_IN_CHAT, INVALID_FILE, INVALID_CONTEXT_MESSAGE }
+    enum class Reason { USER_NOT_IN_CHAT, INVALID_FILE, INVALID_CONTEXT_MESSAGE, MUST_BE_ADMIN }
 }
 
 data class InvalidPicMessage(val reason: Reason) {
-    enum class Reason { USER_NOT_IN_CHAT, INVALID_FILE, INVALID_CONTEXT_MESSAGE, INVALID_CAPTION }
+    enum class Reason { USER_NOT_IN_CHAT, INVALID_FILE, INVALID_CONTEXT_MESSAGE, INVALID_CAPTION, MUST_BE_ADMIN }
 }
 
 data class PicMessageRequest(